+++ src/addon_generator/config/__init__.py
@@ -1,0 +1,9 @@
+from .models import MappingConfigModel
+from .schema_validator import MappingConfigError, load_mapping_config_model, parse_mapping_config_dict
+
+__all__ = [
+    "MappingConfigError",
+    "MappingConfigModel",
+    "load_mapping_config_model",
+    "parse_mapping_config_dict",
+]